pyproject.toml
[project]
name = "PyQWebWindow"
version = "0.6.5"
authors = [
  { name="BHznJNs", email="441768875@qq.com" },
]
description = "Another way to build Python webview GUI applications."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]
dependencies = [
  "PySide6==6.9.0",
  "pyzmq==26.4.0",
]

[project.urls]
Homepage = "https://github.com/BHznJNs/PyQWebWindow"
Issues = "https://github.com/BHznJNs/PyQWebWindow/issues"
